projetos/Parte_2/b04_ContaBancaria_MockWebMVC/src/main/java/com/factory/contabancaria/model/factory/ContaFactory.java
package com.factory.contabancaria.model.factory;

import org.springframework.stereotype.Component;

@Component

public class ContaFactory {

    public CalculoConta tipoServicoConta(String tipo_servico) {
        {
            if (tipo_servico.equalsIgnoreCase("deposito")) {
                return new CalculoDeposito();

            } else if (tipo_servico.equalsIgnoreCase("saque")) {
                return new CalculoSaque();

            } else {
                throw new RuntimeException("Tipo de serviço de conta inválido! " + tipo_servico);
            }
        }

    }
}

